# exportLog() now takes a const reference, yielding some speedup