Before this change:
```
ETHEREUM_TEST_PATH="../../tests" test/testeth -t "StateTestsGeneral"  86,34s user 1,35s system 74% cpu 1:58,06 total
```
After this change:
```
ETHEREUM_TEST_PATH="../../tests" test/testeth -t "StateTestsGeneral"  67,13s user 1,12s system 98% cpu 1:09,04 total
```