@pirapira 2017-07-10
. @fjl was profiling `testeth`, and he told me this function is super slow.