@fjl 2017-07-11
I'm pretty sure that this is not why it's slow ;). Still nicer to pass reference though.