@pirapira 2017-07-11
@fjl yeah, I'm still looking for something like "opening the same file 100 times as the loop loops".